libutil: blobref_validate_hashtype returns size
Problem: the size of the hash digest corresponding to a blobref
is not disclosed to users of the blobref class.

Have blobref_validate_hashtype() return the digest size on
success, instead of zero.

Update tests.

+++ src/common/libutil/blobref.h
@@ -43,9 +43,10 @@ int blobref_hash (const char *hashtype,
 int blobref_validate (const char *blobref);
 
 /* Check the validity of hash type (by name)
+ * If valid, the digest size is returned.
+ * If invalid, -1 is returned with errno set.
  */
-int blobref_validate_hashtype (const char *name);
-
+ssize_t blobref_validate_hashtype (const char *name);
 
 #endif /* _UTIL_BLOBREF_H */
 /*
